Revert some changes in existing files to avoid merge conflicts.
