+++ packages/core/src/core/mongo-db.ts
@@ -1,10 +1,33 @@
-import { MongoClient, Collection } from "mongodb";
-import type {
-    HandlerRole,
-    OrchestratorChat,
-    OrchestratorMessage,
-    ScheduledTask,
-} from "./types";
+import { MongoClient, Collection, ObjectId } from "mongodb";
+import type { HandlerRole } from "./types";
+
+// Define the shape of a scheduled task document in Mongo
+export interface ScheduledTask {
+    _id?: ObjectId;
+    userId: string;
+    handlerName: string; // Which IOHandler to invoke
+    taskData: Record<string, any>; // Arbitrary data passed to the handler
+    nextRunAt: Date; // When the task is next due
+    intervalMs?: number; // If present, re-schedule after each run
+    status: "pending" | "running" | "completed" | "failed";
+    createdAt: Date;
+    updatedAt: Date;
+}
+
+export interface OrchestratorMessage {
+    role: HandlerRole; // "input" | "output" | "action"
+    name: string; // The IOHandler name
+    data: unknown; // Arbitrary data your orchestrator is processing
+    timestamp: Date;
+}
+
+export interface OrchestratorChat {
+    _id?: ObjectId;
+    userId: string;
+    createdAt: Date;
+    updatedAt: Date;
+    messages: OrchestratorMessage[];
+}
 
 export class MongoDb {
     private client: MongoClient;
@@ -71,7 +94,7 @@ export class MongoDb {
         taskData: Record<string, any> = {},
         nextRunAt: Date,
         intervalMs?: number
-    ): Promise<string> {
+    ): Promise<ObjectId> {
         const now = new Date();
         const doc: ScheduledTask = {
             userId,
@@ -112,7 +135,7 @@ export class MongoDb {
     /**
      * Marks a task's status as "running". Typically called right before invoking it.
      */
-    public async markRunning(taskId: string): Promise<void> {
+    public async markRunning(taskId: ObjectId): Promise<void> {
         const now = new Date();
         await this.collection.updateOne(
             { _id: taskId },
@@ -128,7 +151,10 @@ export class MongoDb {
     /**
      * Marks a task as completed (or failed).
      */
-    public async markCompleted(taskId: string, failed = false): Promise<void> {
+    public async markCompleted(
+        taskId: ObjectId,
+        failed = false
+    ): Promise<void> {
         const now = new Date();
         await this.collection.updateOne(
             { _id: taskId },
@@ -145,7 +171,7 @@ export class MongoDb {
      * Updates a task to run again in the future (if intervalMs is present).
      */
     public async updateNextRun(
-        taskId: string,
+        taskId: ObjectId,
         newRunTime: Date
     ): Promise<void> {
         const now = new Date();
@@ -187,7 +213,7 @@ export class MongoDb {
      * Creates a new "orchestrator" document for a user, returning its generated _id.
      * This can represent a "new chat/session" with the agent.
      */
-    public async createOrchestrator(userId: string): Promise<string> {
+    public async createOrchestrator(userId: string): Promise<ObjectId> {
         const chat: OrchestratorChat = {
             userId: userId,
             createdAt: new Date(),
@@ -207,7 +233,7 @@ export class MongoDb {
      * @param data - The data payload to store (e.g., text, JSON from APIs, etc).
      */
     public async addMessage(
-        orchestratorId: string,
+        orchestratorId: ObjectId,
         role: HandlerRole,
         name: string,
         data: unknown
@@ -234,7 +260,7 @@ export class MongoDb {
      * Retrieves all messages in a specific orchestrator's conversation.
      */
     public async getMessages(
-        orchestratorId: string
+        orchestratorId: ObjectId
     ): Promise<OrchestratorMessage[]> {
         const doc = await this.orchestratorCollection.findOne({
             _id: orchestratorId,
@@ -256,7 +282,7 @@ export class MongoDb {
      * Retrieves a single orchestrator document by its ObjectId.
      */
     public async getOrchestratorById(
-        orchestratorId: string
+        orchestratorId: ObjectId
     ): Promise<OrchestratorChat | null> {
         return this.orchestratorCollection.findOne({ _id: orchestratorId });
     }
